refactor: move error-boundary into screen-components
