+++ src/modules/fare-contracts/FareContractOrReservation.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import {Reservation} from '@atb/ticketing';
 import {FareContractType} from '@atb-as/utils';
 import {TicketingTexts, useTranslation} from '@atb/translations';
-import {ErrorBoundary} from '@atb/error-boundary';
+import {ErrorBoundary} from '@atb/screen-components/error-boundary';
 import {FareContractView} from './FareContractView';
 
 export function FareContractOrReservation({
